server/answer_queue: Review material based on source, not stage
Before we were reviewing ug material that was part of our current stage, but
that means that classes sit in an isolated world to any open-access versions.

The downside of this is we now have to have a template per-stage, but that also
gives us places to hang e.g. output tags, which were a problem before.